feat: split integration tests for lazy oracle
